Search refactored, Profile Started
