+++ client/src/App.js
@@ -5,7 +5,7 @@ import "./index.css";
 import "bootstrap/dist/css/bootstrap.min.css";
 import Navbar from './components/Navbar';
 import Home from './pages/Home';
-
+import Profile from './pages/Profile';
 import {
   // ApolloClient,
   createHttpLink,
@@ -43,6 +43,7 @@ function App() {
           <Navbar />
           <Switch>
             <Route exact path="/" component={Home} />
+            <Route exact path="/Profile" component={Profile} />
           </Switch>
         </>
       </Router>
