+++ client/src/pages/Profile.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import { GET_MOVIES } from "../utils/queries"
+import { gql, useQuery } from '@apollo/client';
+
+
+const Profile = () => {
+    return (
+        <h2>Connected Test</h2>
+    )
+}
+
+export default Profile;
